+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
 ## Version 1.0
-
+- (B.3) Added weights and biases to Feed forward neural network
 - (B.2) _CELU_ implementation.
 - (B.1) _Swish_ activation function.